refactor(player): move `player` to separate file
